Move the sqfvm wrapper to a separate file
